Fixup versions introduce asdf
Why:
We would like to start the process of updating this repo's libraries to
up to date versions. Currently the app doesn't run on a clean install so
we need to update a few existing packages before introducing major
version updates in Rails and React.

This commit:
Adds asdf and updates a few gems / packages to get the app running
locally.